requirements.txt
pyautogen>=0.2.0
requests>=2.31.0
python-owasp-zap-v2.4>=0.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
ollama>=0.1.4 